pyproject.toml
[tool.poetry]
name = "dio-bank"
version = "0.1.0"
description = ""
authors = ["Juliano <juliano.co.lopes@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "*"
flask-sqlalchemy = "*"
flask-migrate = "*"
flask-jwt-extended = "*"
psycopg2-binary = "^2.9.9"
gunicorn = "^22.0.0"
flask-bcrypt = "*"
flask-marshmallow = "*"
marshmallow-sqlalchemy = "*"
apispec-webframeworks = "*"

[tool.poetry.group.dev.dependencies]
pytest = "*"
pytest-mock = "*"

[tool.pytest.ini_options]
addopts = "-ra -vv"
testpaths = ["tests"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
